Refactor BackFlow measurement functions into dedicated source
