+++ src/mVMC/calgrn.c
@@ -177,96 +177,5 @@ void CalculateGreenFunc(const double w, const double complex ip, int *eleIdx, in
 }
 
 
-void CalculateGreenFuncBF(const double w, const double complex ip, int *eleIdx, int *eleCfg,
-                          int *eleNum, int *eleProjCnt, const int *eleProjBFCnt) {
 
-  int idx,idx0,idx1;
-  int ri,rj,s,rk,rl,t;
-  double complex tmp;
-  int *myEleIdx, *myEleNum, *myEleCfg, *myProjCntNew, *myProjBFCntNew;
-  double complex* mySltBFTmp;
-  double complex* myBuffer;
-
-  RequestWorkSpaceThreadInt(Nsize+2*Nsite2+NProj+16*Nsite*Nrange);
-  RequestWorkSpaceThreadComplex(NQPFull+2*Nsize+NQPFull*Nsite2*Nsite2);
-  /* GreenFunc1: NQPFull, GreenFunc2: NQPFull+2*Nsize */
-
-#pragma omp parallel default(shared)\
-  private(myEleIdx,myEleNum,myEleCfg,myProjCntNew,myProjBFCntNew,myBuffer,mySltBFTmp)
-  {
-    myEleIdx = GetWorkSpaceThreadInt(Nsize);
-    myEleNum = GetWorkSpaceThreadInt(Nsite2);
-    myEleCfg = GetWorkSpaceThreadInt(Nsite2);
-    myProjCntNew   = GetWorkSpaceThreadInt(NProj);
-    myProjBFCntNew = GetWorkSpaceThreadInt(16*Nsite*Nrange);
-    myBuffer   = GetWorkSpaceThreadComplex(NQPFull+2*Nsize);
-    mySltBFTmp = GetWorkSpaceThreadComplex(NQPFull*Nsite2*Nsite2);
-
-#pragma loop noalias
-    for(idx=0;idx<Nsize;idx++) myEleIdx[idx] = eleIdx[idx];
-#pragma loop noalias
-    for(idx=0;idx<Nsite2;idx++) myEleNum[idx] = eleNum[idx];
-#pragma loop noalias
-    for(idx=0;idx<Nsite2;idx++) myEleCfg[idx] = eleCfg[idx];
-
-    StoreSlaterElmBF_fcmp(mySltBFTmp);
-#pragma omp master
-    {StartTimer(50);}
-
-#pragma omp for private(idx,ri,rj,s,tmp) schedule(dynamic) nowait
-    for(idx=0;idx<NCisAjs;idx++) {
-      ri = CisAjsIdx[idx][0];
-      rj = CisAjsIdx[idx][2];
-      s  = CisAjsIdx[idx][3];
-      tmp = GreenFunc1BF(ri,rj,s,ip,mySltBFTmp,myEleIdx,myEleCfg,myEleNum,eleProjCnt,myProjCntNew,eleProjBFCnt,myProjBFCntNew,myBuffer);
-      LocalCisAjs[idx] = tmp;
-    }
-
-#pragma omp master
-    {StopTimer(50);StartTimer(51);}
-
-#pragma omp for private(idx,ri,rj,s,rk,rl,t,tmp) schedule(dynamic)
-    for(idx=0;idx<NCisAjsCktAltDC;idx++) {
-      ri = CisAjsCktAltDCIdx[idx][0];
-      rj = CisAjsCktAltDCIdx[idx][2];
-      s  = CisAjsCktAltDCIdx[idx][1];
-      rk = CisAjsCktAltDCIdx[idx][4];
-      rl = CisAjsCktAltDCIdx[idx][6];
-      t  = CisAjsCktAltDCIdx[idx][5];
-      tmp = GreenFunc2BF(ri,rj,rk,rl,s,t,ip,mySltBFTmp,myEleIdx,myEleCfg,myEleNum,eleProjCnt,
-                         myProjCntNew,eleProjBFCnt,myProjBFCntNew,myBuffer);
-      LocalCisAjsCktAltDC[idx] = tmp;
-    }
-
-#pragma omp master
-    {StopTimer(51);StartTimer(52);}
-
-#pragma omp for private(idx) nowait
-    for(idx=0;idx<NCisAjs;idx++) {
-      PhysCisAjs[idx] += w*LocalCisAjs[idx];
-    }
-
-#pragma omp for private(idx) nowait
-    for (idx=0;idx<NCisAjsCktAltDC;idx++) {
-      PhysCisAjsCktAltDC[idx] += w*LocalCisAjsCktAltDC[idx];
-    }
-
-#pragma omp master
-    {StopTimer(52);StartTimer(53);}
-
-#pragma omp for private(idx,idx0,idx1) nowait
-    for(idx=0;idx<NCisAjsCktAlt;idx++) {
-      idx0 = CisAjsCktAltIdx[idx][0];
-      idx1 = CisAjsCktAltIdx[idx][1];
-      PhysCisAjsCktAlt[idx] += w*LocalCisAjs[idx0]*conj(LocalCisAjs[idx1]);
-    }
-
-#pragma omp master
-    {StopTimer(53);}
-  }
-
-  ReleaseWorkSpaceThreadInt();
-  ReleaseWorkSpaceThreadComplex();
-  return;
-}
 #endif
