+++ src/mVMC/calham.c
@@ -403,149 +403,3 @@ double complex CalculateHamiltonian2(const double complex ip, int *eleIdx, const
   ReleaseWorkSpaceThreadComplex();
   return e;
 }
-
-double complex CalculateHamiltonianBF_fcmp(const double complex ip, int *eleIdx, const int *eleCfg,
-                              int *eleNum, const int *eleProjCnt, const int *eleProjBFCnt)
-{
-  const int *n0 = eleNum;
-  const int *n1 = eleNum + Nsite;
-  double complex e=0.0, tmp;
-  int idx;
-  int ri,rj,s,rk,rl,t;
-  int *myEleIdx, *myEleNum, *myEleCfg, *myProjCntNew, *myProjBFCntNew;
-  //double sltTmp[NThread*NQPFull*Nsite2*Nsite2];
-  double complex *mySltBFTmp;
-  double complex *myBuffer;
-  double complex myEnergy;
-
-  RequestWorkSpaceThreadInt(Nsize+2*Nsite2+NProj+16*Nsite*Nrange);
-  RequestWorkSpaceThreadComplex(NQPFull+2*Nsize+NQPFull*Nsite2*Nsite2);
-  /* GreenFunc1: NQPFull, GreenFunc2: NQPFull+2*Nsize */
-
-#pragma omp parallel default(shared)\
-  private(myEleIdx,myEleNum,myEleCfg,myProjCntNew,myProjBFCntNew,myBuffer,mySltBFTmp,myEnergy)\
-  reduction(+:e)
-  {
-    myEleIdx = GetWorkSpaceThreadInt(Nsize);
-    myEleNum = GetWorkSpaceThreadInt(Nsite2);
-    myEleCfg = GetWorkSpaceThreadInt(Nsite2);
-    myProjCntNew   = GetWorkSpaceThreadInt(NProj);
-    myProjBFCntNew = GetWorkSpaceThreadInt(16*Nsite*Nrange);
-    myBuffer   = GetWorkSpaceThreadComplex(NQPFull+2*Nsize);
-    mySltBFTmp = GetWorkSpaceThreadComplex(NQPFull*Nsite2*Nsite2);
-
-#pragma loop noalias
-    for(idx=0;idx<Nsize;idx++) myEleIdx[idx] = eleIdx[idx];
-#pragma loop noalias
-    for(idx=0;idx<Nsite2;idx++) myEleNum[idx] = eleNum[idx];
-#pragma loop noalias
-    for(idx=0;idx<Nsite2;idx++) myEleCfg[idx] = eleCfg[idx];
-
-    StoreSlaterElmBF_fcmp(mySltBFTmp);
-#pragma omp barrier
-
-
-    myEnergy = 0.0;
-
-#pragma omp master
-    {StartTimer(70);}
-
-    /* CoulombIntra */
-#pragma omp for private(idx,ri) nowait
-    for(idx=0;idx<NCoulombIntra;idx++) {
-      ri = CoulombIntra[idx];
-      myEnergy += ParaCoulombIntra[idx] * n0[ri] * n1[ri];
-    }
-
-    /* CoulombInter */
-#pragma omp for private(idx,ri,rj) nowait
-    for(idx=0;idx<NCoulombInter;idx++) {
-      ri = CoulombInter[idx][0];
-      rj = CoulombInter[idx][1];
-      myEnergy += ParaCoulombInter[idx] * (n0[ri]+n1[ri]) * (n0[rj]+n1[rj]);
-    }
-
-    /* HundCoupling */
-#pragma omp for private(idx,ri,rj) nowait
-    for(idx=0;idx<NHundCoupling;idx++) {
-      ri = HundCoupling[idx][0];
-      rj = HundCoupling[idx][1];
-      myEnergy -= ParaHundCoupling[idx] * (n0[ri]*n0[rj] + n1[ri]*n1[rj]);
-      /* Caution: negative sign */
-    }
-
-#pragma omp master
-    {StopTimer(70);StartTimer(71);}
-
-    /* Transfer */
-#pragma omp for private(idx,ri,rj,s) schedule(dynamic) nowait
-    for(idx=0;idx<NTransfer;idx++) {
-      ri = Transfer[idx][0];
-      rj = Transfer[idx][2];
-      s  = Transfer[idx][3];
-
-      myEnergy -= ParaTransfer[idx]
-                  //* GreenFunc1(ri,rj,s,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,myBuffer);
-                  *GreenFunc1BF(ri,rj,s,ip,mySltBFTmp,myEleIdx,myEleCfg,myEleNum,eleProjCnt,myProjCntNew,eleProjBFCnt,myProjBFCntNew,myBuffer);
-      /* Caution: negative sign */
-    }
-
-#pragma omp master
-    {StopTimer(71);StartTimer(72);}
-
-    /* Pair Hopping */
-#pragma omp for private(idx,ri,rj) schedule(dynamic) nowait
-    for(idx=0;idx<NPairHopping;idx++) {
-      ri = PairHopping[idx][0];
-      rj = PairHopping[idx][1];
-
-      if (!FlagRBM) {
-        const double complex *rbmCnt;
-        double complex *rbmCntNew;
-        myEnergy += ParaPairHopping[idx]
-                    * GreenFunc2(ri,rj,ri,rj,0,1,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,rbmCnt,rbmCntNew,myBuffer);
-      }
-    }
-
-    /* Exchange Coupling */
-#pragma omp for private(idx,ri,rj,tmp) schedule(dynamic) nowait
-    for(idx=0;idx<NExchangeCoupling;idx++) {
-      ri = ExchangeCoupling[idx][0];
-      rj = ExchangeCoupling[idx][1];
-
-      if (!FlagRBM) {
-        const double complex *rbmCnt;
-        double complex *rbmCntNew;
-        tmp =  GreenFunc2(ri,rj,rj,ri,0,1,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,rbmCnt,rbmCntNew,myBuffer);
-        tmp += GreenFunc2(ri,rj,rj,ri,1,0,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,rbmCnt,rbmCntNew,myBuffer);
-        myEnergy += ParaExchangeCoupling[idx] * tmp;
-      }
-    }
-
-    /* Inter All */
-#pragma omp for private(idx,ri,rj,s,rk,rl,t) schedule(dynamic) nowait
-    for(idx=0;idx<NInterAll;idx++) {
-      ri = InterAll[idx][0];
-      rj = InterAll[idx][2];
-      s  = InterAll[idx][3];
-      rk = InterAll[idx][4];
-      rl = InterAll[idx][6];
-      t  = InterAll[idx][7];
-      if (!FlagRBM) {
-        const double complex *rbmCnt;
-        double complex *rbmCntNew;
-        myEnergy += ParaInterAll[idx]
-                    * GreenFunc2(ri,rj,rk,rl,s,t,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,rbmCnt,rbmCntNew,myBuffer);
-      }
-    }
-
-#pragma omp master
-    {StopTimer(72);}
-
-    e += myEnergy;
-  }
-
-  ReleaseWorkSpaceThreadInt();
-  ReleaseWorkSpaceThreadComplex();
-  return e;
-}
