+++ src/mVMC/calham_real.c
@@ -33,9 +33,6 @@ along with this program. If not, see http://www.gnu.org/licenses/.
 #include "locgrn_real.h"
 #include "calham_real.h"
 
-#ifndef BF_TRANSFER_BATCH_SIZE
-#define BF_TRANSFER_BATCH_SIZE 8
-#endif
 
 ///
 /// \param ip
@@ -213,191 +210,6 @@ double CalculateHamiltonian_real(const double ip, int *eleIdx, const int *eleCfg
   return e;
 }
 
-double CalculateHamiltonianBF_real(const double ip, int *eleIdx, const int *eleCfg,
-                                   int *eleNum, const int *eleProjCnt, const int *eleProjBFCnt) {
-  const int *n0 = eleNum;
-  const int *n1 = eleNum + Nsite;
-  double e=0.0, tmp;
-  int idx;
-  int ri,rj,s,rk,rl,t;
-  int batchNo,batchStart,batchEnd,batchIdx,batchCount,nTransferBatch;
-  int *myEleIdx, *myEleNum, *myEleCfg, *myProjCntNew, *myProjBFCntNew;
-  int *myBatchIdx, *myBatchMsa, *myBatchIcount;
-  //double sltTmp[NThread*NQPFull*Nsite2*Nsite2];
-  double *myBatchVec, *myBatchPfM, *myBatchProj, *myBatchGreen, *myBatchVecStack, *myBatchWStack;
-  double *myBuffer;
-  double myEnergy;
-
-  RequestWorkSpaceThreadInt(Nsize+2*Nsite2+NProj+16*Nsite*Nrange
-                            + BF_TRANSFER_BATCH_SIZE
-                            + BF_TRANSFER_BATCH_SIZE*NQPFull
-                            + BF_TRANSFER_BATCH_SIZE*NQPFull*Nsize);
-  RequestWorkSpaceThreadDouble(NQPFull+2*Nsize
-                               + BF_TRANSFER_BATCH_SIZE*NQPFull*Nsize*Nsize
-                               + BF_TRANSFER_BATCH_SIZE*NQPFull
-                               + 2*BF_TRANSFER_BATCH_SIZE
-                               + 2*BF_TRANSFER_BATCH_SIZE*Nsize*Nsize);
-  nTransferBatch = (NTransfer + BF_TRANSFER_BATCH_SIZE - 1) / BF_TRANSFER_BATCH_SIZE;
-  /* GreenFunc1: NQPFull, GreenFunc2: NQPFull+2*Nsize, BF transfer batch scratch */
-
-#pragma omp parallel default(shared)\
-  private(myEleIdx,myEleNum,myEleCfg,myProjCntNew,myProjBFCntNew,myBuffer,\
-          myBatchIdx,myBatchMsa,myBatchIcount,myBatchVec,myBatchPfM,myBatchProj,myBatchGreen,\
-          myBatchVecStack,myBatchWStack,myEnergy,idx,ri,rj,s,rk,rl,t,tmp,\
-          batchNo,batchStart,batchEnd,batchIdx,batchCount)\
-  reduction(+:e)
-  {
-    myEleIdx = GetWorkSpaceThreadInt(Nsize);
-    myEleNum = GetWorkSpaceThreadInt(Nsite2);
-    myEleCfg = GetWorkSpaceThreadInt(Nsite2);
-    myProjCntNew   = GetWorkSpaceThreadInt(NProj);
-    myProjBFCntNew = GetWorkSpaceThreadInt(16*Nsite*Nrange);
-    myBatchIdx     = GetWorkSpaceThreadInt(BF_TRANSFER_BATCH_SIZE);
-    myBatchIcount  = GetWorkSpaceThreadInt(BF_TRANSFER_BATCH_SIZE*NQPFull);
-    myBatchMsa     = GetWorkSpaceThreadInt(BF_TRANSFER_BATCH_SIZE*NQPFull*Nsize);
-    myBuffer   = GetWorkSpaceThreadDouble(NQPFull+2*Nsize);
-    myBatchVec = GetWorkSpaceThreadDouble(BF_TRANSFER_BATCH_SIZE*NQPFull*Nsize*Nsize);
-    myBatchPfM = GetWorkSpaceThreadDouble(BF_TRANSFER_BATCH_SIZE*NQPFull);
-    myBatchProj = GetWorkSpaceThreadDouble(BF_TRANSFER_BATCH_SIZE);
-    myBatchGreen = GetWorkSpaceThreadDouble(BF_TRANSFER_BATCH_SIZE);
-    myBatchVecStack = GetWorkSpaceThreadDouble(BF_TRANSFER_BATCH_SIZE*Nsize*Nsize);
-    myBatchWStack = GetWorkSpaceThreadDouble(BF_TRANSFER_BATCH_SIZE*Nsize*Nsize);
-
-#pragma loop noalias
-    for(idx=0;idx<Nsize;idx++) myEleIdx[idx] = eleIdx[idx];
-#pragma loop noalias
-    for(idx=0;idx<Nsite2;idx++) myEleNum[idx] = eleNum[idx];
-#pragma loop noalias
-    for(idx=0;idx<Nsite2;idx++) myEleCfg[idx] = eleCfg[idx];
-
-#pragma omp barrier
-
-
-    myEnergy = 0.0;
-
-#pragma omp master
-    {StartTimer(70);}
-
-    /* CoulombIntra */
-#pragma omp for private(idx,ri) nowait
-    for(idx=0;idx<NCoulombIntra;idx++) {
-      ri = CoulombIntra[idx];
-      myEnergy += ParaCoulombIntra[idx] * n0[ri] * n1[ri];
-    }
-
-    /* CoulombInter */
-#pragma omp for private(idx,ri,rj) nowait
-    for(idx=0;idx<NCoulombInter;idx++) {
-      ri = CoulombInter[idx][0];
-      rj = CoulombInter[idx][1];
-      myEnergy += ParaCoulombInter[idx] * (n0[ri]+n1[ri]) * (n0[rj]+n1[rj]);
-    }
-
-    /* HundCoupling */
-#pragma omp for private(idx,ri,rj) nowait
-    for(idx=0;idx<NHundCoupling;idx++) {
-      ri = HundCoupling[idx][0];
-      rj = HundCoupling[idx][1];
-      myEnergy -= ParaHundCoupling[idx] * (n0[ri]*n0[rj] + n1[ri]*n1[rj]);
-      /* Caution: negative sign */
-    }
-
-#pragma omp master
-    {StopTimer(70);StartTimer(71);}
-
-    /* Transfer */
-#pragma omp for private(idx,ri,rj,s) schedule(dynamic) nowait
-    for(batchNo=0;batchNo<nTransferBatch;batchNo++) {
-      batchStart = batchNo * BF_TRANSFER_BATCH_SIZE;
-      batchEnd = batchStart + BF_TRANSFER_BATCH_SIZE;
-      if(batchEnd > NTransfer) batchEnd = NTransfer;
-      batchCount = 0;
-
-      for(idx=batchStart;idx<batchEnd;idx++) {
-        ri = Transfer[idx][0];
-        rj = Transfer[idx][2];
-        s  = Transfer[idx][3];
-
-        if(GreenFunc1BF_real_prepare(ri,rj,s,myBatchGreen+batchCount,myBatchProj+batchCount,
-                                     myBatchVec + batchCount*NQPFull*Nsize*Nsize,
-                                     myBatchMsa + batchCount*NQPFull*Nsize,
-                                     myBatchIcount + batchCount*NQPFull,
-                                     myEleIdx,myEleCfg,myEleNum,eleProjCnt,myProjCntNew,
-                                     eleProjBFCnt,myProjBFCntNew)) {
-          myBatchIdx[batchCount] = idx;
-          batchCount++;
-        } else {
-          myEnergy -= creal(ParaTransfer[idx]) * myBatchGreen[batchCount];
-        }
-        /* Caution: negative sign */
-      }
-
-      if(batchCount > 0) {
-        GreenFunc1BF_real_finish_batch(batchCount, ip, myBatchProj, myBatchIcount, myBatchMsa,
-                                       myBatchVec, myBatchGreen, myBatchPfM,
-                                       myBatchVecStack, myBatchWStack);
-        for(batchIdx=0;batchIdx<batchCount;batchIdx++) {
-          idx = myBatchIdx[batchIdx];
-          myEnergy -= creal(ParaTransfer[idx]) * myBatchGreen[batchIdx];
-        }
-        /* Caution: negative sign */
-      }
-    }
-
-#pragma omp master
-    {StopTimer(71);StartTimer(72);}
-
-    /* Pair Hopping */
-#pragma omp for private(idx,ri,rj) schedule(dynamic) nowait
-    for(idx=0;idx<NPairHopping;idx++) {
-      ri = PairHopping[idx][0];
-      rj = PairHopping[idx][1];
-
-      myEnergy += ParaPairHopping[idx]
-        * GreenFunc2_real(ri,rj,ri,rj,0,1,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,myBuffer);
-    }
-
-    /* Exchange Coupling */
-#pragma omp for private(idx,ri,rj,tmp) schedule(dynamic) nowait
-    for(idx=0;idx<NExchangeCoupling;idx++) {
-      ri = ExchangeCoupling[idx][0];
-      rj = ExchangeCoupling[idx][1];
-
-      tmp =  GreenFunc2_real(ri,rj,rj,ri,0,1,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,myBuffer);
-      tmp += GreenFunc2_real(ri,rj,rj,ri,1,0,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,myBuffer);
-      myEnergy += ParaExchangeCoupling[idx] * tmp;
-    }
-
-    /* Inter All */
-#pragma omp for private(idx,ri,rj,s,rk,rl,t) schedule(dynamic) nowait
-    for(idx=0;idx<NInterAll;idx++) {
-      ri = InterAll[idx][0];
-      rj = InterAll[idx][2];
-      s  = InterAll[idx][3];
-      rk = InterAll[idx][4];
-      rl = InterAll[idx][6];
-      t  = InterAll[idx][7];
-      myEnergy += creal(ParaInterAll[idx])
-        * GreenFunc2_real(ri,rj,rk,rl,s,t,ip,myEleIdx,eleCfg,myEleNum,eleProjCnt,myProjCntNew,myBuffer);
-    }
-
-#pragma omp master
-    {StopTimer(72);}
-
-    e += myEnergy;
-  }
-
-  ReleaseWorkSpaceThreadInt();
-  ReleaseWorkSpaceThreadDouble();
-  return e;
-}
-
-/* Calculate the CoulombIntra, CoulombInter, Hund terms, */
-/* which can be calculated by number operators. */
-///
-/// \param eleNum [in]
-/// \return myEnergy
-/// \version 1.0
 double CalculateHamiltonian0_real(const int *eleNum)
 {
   const int *n0 = eleNum;
