+++ src/mVMC/locgrn_real.c
@@ -431,372 +431,5 @@ double calculateNewPfMN_real_child(const int qpidx, const int n, const int *msa,
   return sgn * pfaff * PfM_real[qpidx];
 }
 
-static int FindBFHopIndex_real(const int *list, const int count, const int value) {
-  int i;
-  for (i = 0; i < count; i++) {
-    if (list[i] == value) return i;
-  }
-  return -1;
-}
-
-static void CopyBFVecRow_real(double *dstVec, const int dstRow,
-                              const double *srcVec, const int srcRow) {
-  int i;
-  double *dst = dstVec + dstRow*Nsize;
-  const double *src = srcVec + srcRow*Nsize;
-  if (dst == src) return;
-  for (i = 0; i < Nsize; i++) {
-    dst[i] = src[i];
-  }
-}
-
-static int MergeBFHopListsVec_real(const int *left, const int *leftCount, const double *leftVec,
-                                   const int *right, const int *rightCount, const double *rightVec,
-                                   int *merged, int *mergedCount, double *mergedVec) {
-  int qpidx, i, outIdx;
-  for (qpidx = 0; qpidx < NQPFull; qpidx++) {
-    int *dst = merged + qpidx * Nsize;
-    const int *leftQp = left + qpidx * Nsize;
-    const int *rightQp = right + qpidx * Nsize;
-    const double *leftVecQp = leftVec + qpidx * Nsize * Nsize;
-    const double *rightVecQp = rightVec + qpidx * Nsize * Nsize;
-    double *mergedVecQp = mergedVec + qpidx * Nsize * Nsize;
-    mergedCount[qpidx] = 0;
-
-    for (i = 0; i < leftCount[qpidx]; i++) {
-      if (FindBFHopIndex_real(dst, mergedCount[qpidx], leftQp[i]) >= 0) continue;
-      if (mergedCount[qpidx] >= Nsize) return 1;
-      outIdx = mergedCount[qpidx];
-      dst[outIdx] = leftQp[i];
-      CopyBFVecRow_real(mergedVecQp, outIdx, leftVecQp, i);
-      mergedCount[qpidx]++;
-    }
-    for (i = 0; i < rightCount[qpidx]; i++) {
-      if (FindBFHopIndex_real(dst, mergedCount[qpidx], rightQp[i]) >= 0) continue;
-      if (mergedCount[qpidx] >= Nsize) return 1;
-      outIdx = mergedCount[qpidx];
-      dst[outIdx] = rightQp[i];
-      CopyBFVecRow_real(mergedVecQp, outIdx, rightVecQp, i);
-      mergedCount[qpidx]++;
-    }
-  }
-  return 0;
-}
-
-
-/* Calculate 1-body Green function <CisAjs> */
-/* buffer size = NQPFull */
-int GreenFunc1BF_real_prepare(const int ri, const int rj, const int s, double *greenValue,
-                    double *projRatio, double *vecM, int *msaTmp, int *icount,
-                    int *eleIdx, int *eleCfg, int *eleNum, const int *eleProjCnt,
-                    int *projCntNew, const int *eleProjBFCnt,int *projBFCntNew) {
-  int mj,msj,rsi,rsj;
-
-  StartTimer(84);
-  if(ri==rj) {
-    *greenValue = eleNum[ri+s*Nsite];
-    StopTimer(84);
-    return 0;
-  }
-  if(eleNum[ri+s*Nsite]==1 || eleNum[rj+s*Nsite]==0) {
-    *greenValue = 0.0;
-    StopTimer(84);
-    return 0;
-  }
-
-  mj = eleCfg[rj+s*Nsite];
-  msj = mj + s*Ne;
-  rsi = ri + s*Nsite;
-  rsj = rj + s*Nsite;
-
-  StartTimer(86);
-  eleCfg[rsj] = -1;
-  eleCfg[rsi] = mj;
-  eleIdx[msj] = ri;
-  eleNum[rsj] = 0;
-  eleNum[rsi] = 1;
-  UpdateProjCnt(rj, ri, s, projCntNew, eleProjCnt, eleNum);
-  *projRatio = ProjRatio(projCntNew,eleProjCnt);
-  StopTimer(86);
-
-  StartTimer(81);
-  MakeProjBFCnt(projBFCntNew, eleNum);
-  StopTimer(81);
-  StartTimer(82);
-  UpdateSlaterElmBFGrnVec_real(mj, rj, ri, s, eleIdx, eleCfg, eleNum,
-                               eleProjBFCnt, projBFCntNew, msaTmp, icount, vecM);
-  StopTimer(82);
-
-  StartTimer(88);
-  eleCfg[rsj] = mj;
-  eleCfg[rsi] = -1;
-  eleIdx[msj] = rj;
-  eleNum[rsj] = 1;
-  eleNum[rsi] = 0;
-  StopTimer(88);
-
-  StopTimer(84);
-  return 1;
-}
-
-void GreenFunc1BF_real_finish_batch(const int batchSize, const double ip,
-                    const double *projRatio, const int *icount, const int *msaTmp,
-                    const double *vecM, double *greenValue, double *pfMNew,
-                    double *vecStack, double *wStack) {
-  int batchIdx;
-
-  StartTimer(84);
-  StartTimer(83);
-  CalculateNewPfMBFVecBatched_real(batchSize, icount, msaTmp, pfMNew, 0, NQPFull,
-                                   vecM, vecStack, wStack);
-  StopTimer(83);
-  StartTimer(87);
-  for(batchIdx=0;batchIdx<batchSize;batchIdx++) {
-    greenValue[batchIdx] = projRatio[batchIdx]
-      * CalculateIP_real(pfMNew + batchIdx*NQPFull, 0, NQPFull, MPI_COMM_SELF) / ip;
-  }
-  StopTimer(87);
-  StopTimer(84);
-
-  return;
-}
-
-double GreenFunc1BF_real(const int ri, const int rj, const int s, const double ip, double *bufM,
-                    int *eleIdx, int *eleCfg, int *eleNum, const int *eleProjCnt,
-                    int *projCntNew, const int *eleProjBFCnt,int *projBFCntNew, double *buffer) {
-  double z;
-  int mj,msj,rsi,rsj;
-  //double complex bufM[NQPFull*Nsize*Nsize];
-  double *pfMNew_real = buffer; /* NQPFull */
-  int msaTmp[NQPFull*Nsize],icount[NQPFull];
-
-#define RETURN_GREENFUNC1BF_REAL(value) do { z = (value); StopTimer(84); return z; } while (0)
-
-  StartTimer(84);
-  if(ri==rj) RETURN_GREENFUNC1BF_REAL(eleNum[ri+s*Nsite]);
-  if(eleNum[ri+s*Nsite]==1 || eleNum[rj+s*Nsite]==0) RETURN_GREENFUNC1BF_REAL(0.0);
-
-  /* store SlaterElmBF before hopping */
-  mj = eleCfg[rj+s*Nsite];
-  msj = mj + s*Ne;
-  rsi = ri + s*Nsite;
-  rsj = rj + s*Nsite;
-
-  StartTimer(86);
-  /* hopping */
-  eleCfg[rsj] = -1;
-  eleCfg[rsi] = mj;
-  eleIdx[msj] = ri;
-  eleNum[rsj] = 0;
-  eleNum[rsi] = 1;
-  UpdateProjCnt(rj, ri, s, projCntNew, eleProjCnt, eleNum);
-  z = ProjRatio(projCntNew,eleProjCnt);
-  StopTimer(86);
-
-  /* calculate Pfaffian */
-  //printf("1");
-  StartTimer(81);
-  MakeProjBFCnt(projBFCntNew, eleNum);
-  StopTimer(81);
-  StartTimer(82);
-  UpdateSlaterElmBFGrnVec_real(mj, rj, ri, s, eleIdx, eleCfg, eleNum,
-                               eleProjBFCnt, projBFCntNew, msaTmp, icount, bufM);
-  StopTimer(82);
-  StartTimer(83);
-  CalculateNewPfMBFVec_real(icount, msaTmp, pfMNew_real, 0, NQPFull, bufM);
-  StopTimer(83);
-  StartTimer(87);
-  z *= CalculateIP_real(pfMNew_real, 0, NQPFull, MPI_COMM_SELF);
-  StopTimer(87);
-  //printf("1");
-  //MakeSlaterElmBF(eleNum,projBFCntNew);
-  //UpdateSlaterElmBF2(eleNum,projBFCntNew);
-  //printf("1");
-  //CalculateMAllBF_NoThread(eleIdx,0,NQPFull,bufM);
-  //CalculateMAllBF_NoThread(eleIdx,0,NQPFull,SlaterElmBF);
-  //printf("1\n");
-  //z *= CalculateIP(PfM,0,NQPFull,MPI_COMM_SELF);
-
-  StartTimer(88);
-  /* revert hopping */
-  eleCfg[rsj] = mj;
-  eleCfg[rsi] = -1;
-  eleIdx[msj] = rj;
-  eleNum[rsj] = 1;
-  eleNum[rsi] = 0;
-  StopTimer(88);
-
-  //UpdateSlaterElmBFTmp5(mj, ri, rj, s, eleCfg, eleNum, eleProjBFCnt, msaTmp, icount, bufM);
-  //MakeProjBFCnt(projBFCntNew, eleNum);
-  //MakeSlaterElmBF(eleNum,projBFCntNew);
-  //CalculateMAllBF_NoThread(eleIdx,0,NQPFull,bufM);
-  //CalculateMAllBF_NoThread(eleIdx,0,NQPFull,SlaterElmBF);
-
-  RETURN_GREENFUNC1BF_REAL(z/ip);
-#undef RETURN_GREENFUNC1BF_REAL
-}
-
-/* Calculate 2-body Green function with BackFlow. */
-double GreenFunc2BF_real(const int ri, const int rj, const int rk, const int rl,
-                    const int s, const int t, const double ip, double *bufM,
-                    int *eleIdx, int *eleCfg, int *eleNum, const int *eleProjCnt,
-                    int *projCntNew, const int *eleProjBFCnt,int *projBFCntNew, double *buffer) {
-  double z;
-  int mj,msj,ml,mtl;
-  int rsi,rsj,rtk,rtl;
-  double *pfMNew_real = buffer; /* [NQPFull] */
-  int msaTmp0[NQPFull*Nsize], msaTmp1[NQPFull*Nsize], msaTmp[NQPFull*Nsize];
-  int icount0[NQPFull], icount1[NQPFull], icount[NQPFull];
-  double *vecTmp0 = NULL;
-  size_t vecTmpSize;
-
-#define RETURN_GREENFUNC2BF_REAL(value) do { z = (value); StopTimer(85); return z; } while (0)
-
-  StartTimer(85);
-
-  rsi = ri + s*Nsite;
-  rsj = rj + s*Nsite;
-  rtk = rk + t*Nsite;
-  rtl = rl + t*Nsite;
-
-  if(s==t) {
-    if(rk==rl) {
-      if(eleNum[rtk]==0) RETURN_GREENFUNC2BF_REAL(0.0);
-      else RETURN_GREENFUNC2BF_REAL(GreenFunc1BF_real(ri,rj,s,ip,bufM,eleIdx,eleCfg,eleNum,
-                                    eleProjCnt,projCntNew,eleProjBFCnt,projBFCntNew,buffer));
-    }else if(rj==rl) {
-      RETURN_GREENFUNC2BF_REAL(0.0);
-    }else if(ri==rl) {
-      if(eleNum[rsi]==0) RETURN_GREENFUNC2BF_REAL(0.0);
-      else if(rj==rk) RETURN_GREENFUNC2BF_REAL(1.0-eleNum[rsj]);
-      else RETURN_GREENFUNC2BF_REAL(-GreenFunc1BF_real(rk,rj,s,ip,bufM,eleIdx,eleCfg,eleNum,
-                                     eleProjCnt,projCntNew,eleProjBFCnt,projBFCntNew,buffer));
-    }else if(rj==rk) {
-      if(eleNum[rsj]==1) RETURN_GREENFUNC2BF_REAL(0.0);
-      else if(ri==rl) RETURN_GREENFUNC2BF_REAL(eleNum[rsi]);
-      else RETURN_GREENFUNC2BF_REAL(GreenFunc1BF_real(ri,rl,s,ip,bufM,eleIdx,eleCfg,eleNum,
-                                    eleProjCnt,projCntNew,eleProjBFCnt,projBFCntNew,buffer));
-    }else if(ri==rk) {
-      RETURN_GREENFUNC2BF_REAL(0.0);
-    }else if(ri==rj) {
-      if(eleNum[rsi]==0) RETURN_GREENFUNC2BF_REAL(0.0);
-      else RETURN_GREENFUNC2BF_REAL(GreenFunc1BF_real(rk,rl,s,ip,bufM,eleIdx,eleCfg,eleNum,
-                                    eleProjCnt,projCntNew,eleProjBFCnt,projBFCntNew,buffer));
-    }
-  }else{
-    if(rk==rl) {
-      if(eleNum[rtk]==0) RETURN_GREENFUNC2BF_REAL(0.0);
-      else if(ri==rj) RETURN_GREENFUNC2BF_REAL(eleNum[rsi]);
-      else RETURN_GREENFUNC2BF_REAL(GreenFunc1BF_real(ri,rj,s,ip,bufM,eleIdx,eleCfg,eleNum,
-                                    eleProjCnt,projCntNew,eleProjBFCnt,projBFCntNew,buffer));
-    }else if(ri==rj) {
-      if(eleNum[rsi]==0) RETURN_GREENFUNC2BF_REAL(0.0);
-      else RETURN_GREENFUNC2BF_REAL(GreenFunc1BF_real(rk,rl,t,ip,bufM,eleIdx,eleCfg,eleNum,
-                                    eleProjCnt,projCntNew,eleProjBFCnt,projBFCntNew,buffer));
-    }
-  }
-
-  if(eleNum[rsi]==1 || eleNum[rsj]==0 || eleNum[rtk]==1 || eleNum[rtl]==0) RETURN_GREENFUNC2BF_REAL(0.0);
-
-  mj = eleCfg[rsj];
-  ml = eleCfg[rtl];
-  msj = mj + s*Ne;
-  mtl = ml + t*Ne;
-
-  if(NQPFull <= 0 || Nsize <= 0) RETURN_GREENFUNC2BF_REAL(0.0);
-  vecTmpSize = (size_t)NQPFull * (size_t)Nsize * (size_t)Nsize;
-  vecTmp0 = (double *)malloc(sizeof(double)*vecTmpSize);
-  if(vecTmp0 == NULL) RETURN_GREENFUNC2BF_REAL(0.0);
-
-  StartTimer(86);
-  eleCfg[rtl] = -1;
-  eleCfg[rtk] = ml;
-  eleIdx[mtl] = rk;
-  eleNum[rtl] = 0;
-  eleNum[rtk] = 1;
-  UpdateProjCnt(rl, rk, t, projCntNew, eleProjCnt, eleNum);
-
-  eleCfg[rsj] = -1;
-  eleCfg[rsi] = mj;
-  eleIdx[msj] = ri;
-  eleNum[rsj] = 0;
-  eleNum[rsi] = 1;
-  UpdateProjCnt(rj, ri, s, projCntNew, projCntNew, eleNum);
-
-  if(!IsSectorStateAllowed(eleNum)) {
-    eleCfg[rtl] = ml;
-    eleCfg[rtk] = -1;
-    eleIdx[mtl] = rl;
-    eleNum[rtl] = 1;
-    eleNum[rtk] = 0;
-    eleCfg[rsj] = mj;
-    eleCfg[rsi] = -1;
-    eleIdx[msj] = rj;
-    eleNum[rsj] = 1;
-    eleNum[rsi] = 0;
-    StopTimer(86);
-    free(vecTmp0);
-    RETURN_GREENFUNC2BF_REAL(0.0);
-  }
-
-  z = ProjRatio(projCntNew,eleProjCnt);
-  StopTimer(86);
-
-  StartTimer(81);
-  MakeProjBFCnt(projBFCntNew, eleNum);
-  StopTimer(81);
-  StartTimer(82);
-  UpdateSlaterElmBFGrnVec_real(ml, rl, rk, t, eleIdx, eleCfg, eleNum,
-                               eleProjBFCnt, projBFCntNew, msaTmp0, icount0, vecTmp0);
-  UpdateSlaterElmBFGrnVec_real(mj, rj, ri, s, eleIdx, eleCfg, eleNum,
-                               eleProjBFCnt, projBFCntNew, msaTmp1, icount1, bufM);
-  StopTimer(82);
-
-  StartTimer(90);
-  if (MergeBFHopListsVec_real(msaTmp0, icount0, vecTmp0, msaTmp1, icount1, bufM, msaTmp, icount, vecTmp0) != 0) {
-    StopTimer(90);
-    StartTimer(88);
-    eleCfg[rtl] = ml;
-    eleCfg[rtk] = -1;
-    eleIdx[mtl] = rl;
-    eleNum[rtl] = 1;
-    eleNum[rtk] = 0;
-    eleCfg[rsj] = mj;
-    eleCfg[rsi] = -1;
-    eleIdx[msj] = rj;
-    eleNum[rsj] = 1;
-    eleNum[rsi] = 0;
-    StopTimer(88);
-    free(vecTmp0);
-    RETURN_GREENFUNC2BF_REAL(0.0);
-  }
-  StopTimer(90);
-
-  StartTimer(83);
-  CalculateNewPfMBFVecWithStride_real(icount, msaTmp, Nsize, pfMNew_real, 0, NQPFull, vecTmp0, Nsize);
-  StopTimer(83);
-  StartTimer(87);
-  z *= CalculateIP_real(pfMNew_real, 0, NQPFull, MPI_COMM_SELF);
-  StopTimer(87);
-
-  StartTimer(88);
-  eleCfg[rtl] = ml;
-  eleCfg[rtk] = -1;
-  eleIdx[mtl] = rl;
-  eleNum[rtl] = 1;
-  eleNum[rtk] = 0;
-  eleCfg[rsj] = mj;
-  eleCfg[rsi] = -1;
-  eleIdx[msj] = rj;
-  eleNum[rsj] = 1;
-  eleNum[rsi] = 0;
-  StopTimer(88);
-
-  free(vecTmp0);
-
-  RETURN_GREENFUNC2BF_REAL(z/ip);
-#undef RETURN_GREENFUNC2BF_REAL
-}
-
 
 #endif
